intel bsp .cfg files: remove redundant sound configs
The sound configs are defined in features/sound/snd_hda_intel.cfg.
If some of these configs are defined differently like kernel builtin
instead of modules, then it changes the timing of driver loading causing
audio breakages for some of the BSPs.

This commit fixes the audio issues with some of the Intel BSPs
with the v3.14 kernel.

Signed-off-by: Nitin A Kamble <nitin.a.kamble@intel.com>
Signed-off-by: Bruce Ashfield <bruce.ashfield@windriver.com>