rcu: Add const annotation to char * for RCU tracepoints and functions
Issue: LIN6-7388

1/1 [
Author: Linhao Zhang
Email: linhao.zhang@windriver.com
Subject: rcu: Add const annotation to char * for RCU tracepoints and functions
Date: Fri, 16 May 2014 16:49:08 +0800

commit e66c33d579ea566d10e8c8695a7168aae3e02992 upstream.

All the RCU tracepoints and functions that reference char pointers do
so with just 'char *' even though they do not modify the contents of
the string itself. This will cause warnings if a const char * is used
in one of these functions.

The RCU tracepoints store the pointer to the string to refer back to them
when the trace output is displayed. As this can be minutes, hours or
even days later, those strings had better be constant.

This change also opens the door to allow the RCU tracepoint strings and
their addresses to be exported so that userspace tracing tools can
translate the contents of the pointers of the RCU tracepoints.

Signed-off-by: Steven Rostedt <rostedt@goodmis.org>
Signed-off-by: Paul Gortmaker <paul.gortmaker@windriver.com>
Signed-off-by: Linhao Zhang <linhao.zhang@windriver.com>
]

Signed-off-by: Yue Tao <yue.tao@windriver.com>