Fix AP powersave TX vs. wakeup race - linux - CVE-2014-2706
Issue: LIN5-18657

1/1 [
Author: Jian Liu
Email: jian.liu@windriver.com
Subject: Fix AP powersave TX vs. wakeup race
Date: Wed, 7 May 2014 16:05:59 +0800

commit 1d147bfa64293b2723c4fec50922168658e613ba upstream

There is a race between the TX path and the STA wakeup:
while a station is sleeping, mac80211 buffers frames
until it wakes up, then the frames are transmitted.
However, the RX and TX path are concurrent, so the
packet indicating wakeup can be processed while a
packet is being transmitted.
This can lead to a situation where the buffered frames
list is emptied on the one side, while a frame is
being added on the other side, as the station is still
seen as sleeping in the TX path.
As a result, the newly added frame will not be send
anytime soon. It might be sent much later
(and out of order) when the station goes to sleep
and wakes up the next time.
Additionally, it can lead to the crash below.
Fix all this by synchronising both paths with a new
lock. Both path are not fastpath since they handle
PS situations.

Reported-by: Yaara Rozenblum <yaara.rozenblum@intel.com>
Reviewed-by: Stanislaw Gruszka <sgruszka@redhat.com>
Signed-off-by: Emmanuel Grumbach <emmanuel.grumbach@intel.com>
Signed-off-by: Jian Liu <jian.liu@windriver.com>
]

Signed-off-by: Yue Tao <yue.tao@windriver.com>