net:fix Memory leak on Marvell Armada XP
Issue: LIN5-18408

1/1 [
Author: Baogen Shang
Email: baogen.shang@windriver.com
Subject: net:fix Memory leak on Marvell Armada XP
Date: Fri, 16 May 2014 11:13:42 +0800

On Marvell Armada XP,If Let the NET_SKB_RECYCLE = y,It will be lead to the
memory leak When the network card receive packages.the root cause is that the
callback function "mv_eth_skb_recycle" responsible for recycling of memory
never be called.But considering the SKB recycling mechanism has been unused,
and dropped from Linux completely since 3.5.x.So the fix is that disable it
in the BSP config, and then patch the kernel and add:

   depends on BROKEN

To the NET_SKB_RECYCLE Kconfig. This let it can't be enabled again.

Signed-off-by: baogen shang <baogen.shang@windriver.com>
]

Signed-off-by: Yue Tao <yue.tao@windriver.com>