FIX: lttng build error if \!CONFIG_UID16 on arm
Issue: LIN6-7472

1/1 [
Author: Linhao Zhang
Email: linhao.zhang@windriver.com
Subject: FIX: lttng build error if !CONFIG_UID16 on arm
Date: Thu, 22 May 2014 09:52:43 +0800

when CONFIG_UID16 is not set, the old_uid_t is not defined.
If the macros such as OVERRIDE_32_sys_getuid16 is not defined,
the type old_uid_t will be used in lttng,
so define the macros to avoid that.

Signed-off-by: Linhao Zhang <linhao.zhang@windriver.com>
]

Signed-off-by: Yue Tao <yue.tao@windriver.com>