Axxia:resolve a dead-lock issue with axxia gic
1/1 [
Author: Jiang Lu
Email: lu.jiang@windriver.com
Subject: Axxia:resolve a dead-lock issue with axxia gic
Date: Wed May 28 03:12:31 2014 +0000

Fixed following calltrace on lsi-axm55xx platform:

BUG: spinlock lockup suspected on CPU#0, swapper/0/0
BUG: spinlock lockup suspected on CPU#8, clock_settime_4/12914
 lock: irq_desc+0x2450/0x10000, .magic: dead4ead, .owner: irqbalance/436, .owner_cpu: 2
CPU: 8 PID: 12914 Comm: clock_settime_4 Not tainted 3.10.39-grsec-WR6.0.0.6_secure #1
[<c0416bd8>] (unwind_backtrace+0x0/0xec) from [<c041221c>] (show_stack+0x20/0x24)
[<c041221c>] (show_stack+0x20/0x24) from [<c0a217a4>] (dump_stack+0x20/0x28)
[<c0a217a4>] (dump_stack+0x20/0x28) from [<c0a21f2c>] (spin_dump+0x80/0x94)
[<c0a21f2c>] (spin_dump+0x80/0x94) from [<c0748858>] (do_raw_spin_lock+0x124/0x170)
[<c0748858>] (do_raw_spin_lock+0x124/0x170) from [<c0a273e0>] (_raw_spin_lock+0x50/0x58)
[<c0a273e0>] (_raw_spin_lock+0x50/0x58) from [<c04a46e4>] (handle_fasteoi_irq+0x24/0x10c)
[<c04a46e4>] (handle_fasteoi_irq+0x24/0x10c) from [<c04a0cac>] (generic_handle_irq+0x30/0x40)
[<c04a0cac>] (generic_handle_irq+0x30/0x40) from [<c040ee84>] (handle_IRQ+0x78/0xa0)
[<c040ee84>] (handle_IRQ+0x78/0xa0) from [<c04087fc>] (axxia_gic_handle_irq+0x174/0x19c)
[<c04087fc>] (axxia_gic_handle_irq+0x174/0x19c) from [<c0a282c4>] (__irq_svc+0x44/0x78)

On axxia platform, there are 16 cores on SOC. The 16 cores are divided
into 4 clusters. Each irq must be assigned to a cluster, the irq
operations must be performed on core within same cluster. IRQ
operations include mask/unmask an irq, change type for an irq, and
update affinity for an irq. So when received an irq operation on cpu
didn't belong to the cluster, kernel need send an IPI to corresponding
cpu to complete the operation, then let the issuer wait.

In kernel, irq operation routine and ISR need acquire the lock of the
irq desc. Since the irq operation routine and ISR may be served on
different cores at the same time. This would introduce a dead lock in
kernel.

This patch force the IPI issuer release the lock before waiting, which
could allow the receiver to acquire the lock to complete the operation.

Signed-off-by: Jiang Lu <lu.jiang@windriver.com>
]

Signed-off-by: Yang Shi <yang.shi@windriver.com>