tasklet: fix a race condition in the state-bit flipping in softirq.c
Issue: LIN6-7505

1/1 [
Author: Fredrik Markstrom
Email: fredrik.markstrom@gmail.com
Subject: tasklet: fix a race condition in the state-bit flipping in softirq.c
Date: Tue, 22 Apr 2014 15:20:04 +0800

If a thread got preempted in tasklet_schedule after setting the SCHED bit but
before taking the "lock", the restart code in tasklet_action() can can run the
tasklet and clear SCHED. If other threads in the system schedule the task
again before the first tasklet_schedule() may run again it will successfully
set SCHED again and add the tasklet to a tasklet queue. If this happens and the
first tasklet_schedule() continues it will successfully get the lock and it
will verify that the SCHED bit is taken and add the tasklet queue, this is
wrong since it's already in a tasklet queue. The consequences are that the
tasklet will run one time too many and tasklets behind it in the first list
potentially on the wrong cpu.

Signed-off-by: Fredrik Markstrom <fredrik.markstrom@gmail.com>
Signed-off-by: baogen shang <baogen.shang@windriver.com>
]

Signed-off-by: Yue Tao <yue.tao@windriver.com>