Btrfs: use right root when checking for hash collision - linux - CVE-2012-5375
Issue: LIN4-23170

1/2 [
Author: Zhang Xiao
Email: xiao.zhang@windriver.com
Subject: Btrfs: fix hash overflow handling
Date: Tue, 10 Jun 2014 17:13:40 +0800

Commit 9c52057c698fb96f8f07e7a4bcf4801a092bda89 from upstream.

The handling for directory crc hash overflows was fairly obscure,
split_leaf returns EOVERFLOW when we try to extend the item and that is
supposed to bubble up to userland.  For a while it did so, but along the
way we added better handling of errors and forced the FS readonly if we
hit IO errors during the directory insertion.

Along the way, we started testing only for EEXIST and the EOVERFLOW case
was dropped.  The end result is that we may force the FS readonly if we
catch a directory hash bucket overflow.

This fixes a few problem spots.  First I add tests for EOVERFLOW in the
places where we can safely just return the error up the chain.

btrfs_rename is harder though, because it tries to insert the new
directory item only after it has already unlinked anything the rename
was going to overwrite.  Rather than adding very complex logic, I added
a helper to test for the hash overflow case early while it is still safe
to bail out.

Snapshot and subvolume creation had a similar problem, so they are using
the new helper now too.

Signed-off-by: Chris Mason <chris.mason@fusionio.com>
Reported-by: Pascal Junod <pascal@junod.info>
Signed-off-by: Zhang Xiao <xiao.zhang@windriver.com>
]

2/2 [
Author: Zhang Xiao
Email: xiao.zhang@windriver.com
Subject: Btrfs: use right root when checking for hash collision
Date: Tue, 10 Jun 2014 17:14:49 +0800

Commit 4871c1588f92c6c13f4713a7009f25f217055807 from upstream.

btrfs_rename was using the root of the old dir instead of the root of the new
dir when checking for a hash collision, so if you tried to move a file into a
subvol it would freak out because it would see the file you are trying to move
in its current root.  This fixes the bug where this would fail

btrfs subvol create test1
btrfs subvol create test2
mv test1 test2.

Thanks to Chris Murphy for catching this,

Cc: stable@vger.kernel.org
Reported-by: Chris Murphy <lists@colorremedies.com>
Signed-off-by: Josef Bacik <jbacik@fusionio.com>
Signed-off-by: Chris Mason <chris.mason@fusionio.com>
Signed-off-by: Zhang Xiao <xiao.zhang@windriver.com>
]

Signed-off-by: Yue Tao <yue.tao@windriver.com>