wrhv: pci: Find the correct parent bus before allocating resource
Issue: LIN4-31440

1/1 [
Author: Xufeng Zhang
Email: xufeng.zhang@windriver.com
Subject: wrhv: pci: Find the correct parent bus before allocating resource
Date: Mon, 23 Jun 2014 10:56:05 +0800

There is case that the root resource descriptor passed to allocate_resource()
is the desired resource's grandparent, when this happens, the allocating request
always failed because of two reasons:
1). We may failed to find a valid memory space(consistent with XML
configuration) in find_resource().
2). The conflict entry would be found in __request_resource().

So before we call allocate_resource(), we should guarantee that we are
on the correct resource tree(parent's tree).

Signed-off-by: Xufeng Zhang <xufeng.zhang@windriver.com>
]

Signed-off-by: Yue Tao <yue.tao@windriver.com>