net: Add variants of capable for use on netlink messages - linux - CVE-2014-0181
Issue: LIN6-7270

1/4 [
Author: Jian Liu
Email: jian.liu@windriver.com
Subject: net: Use netlink_ns_capable to verify the permisions of netlink messages
Date: Tue, 24 Jun 2014 14:55:56 +0800

commit 90f62cf30a78721641e08737bda787552428061e upstream

It is possible by passing a netlink socket to a more privileged executable
and then to fool that executable into writing to the socket data that
happens to be valid netlink message to do something that privileged
executable did not intend to do. To keep this from happening replace bare
capable and ns_capable calls with netlink_capable, netlink_net_calls and
netlink_ns_capable calls. Which act the same as the previous calls except
they verify that the opener of the socket had the desired permissions as well.

Reported-by: Andy Lutomirski <luto@amacapital.net>
Signed-off-by: "Eric W. Biederman" <ebiederm@xmission.com>
Signed-off-by: David S. Miller <davem@davemloft.net>
Signed-off-by: Jian Liu <jian.liu@windriver.com>
]

2/4 [
Author: Jian Liu
Email: jian.liu@windriver.com
Subject: netlink: Rename netlink_capable netlink_allowed
Date: Tue, 24 Jun 2014 15:25:58 +0800

commit 5187cd055b6e81fc6526109456f8b20623148d5f upstream

netlink_capable is a static internal function in af_netlink.c and we
have better uses for the name netlink_capable.

Signed-off-by: "Eric W. Biederman" <ebiederm@xmission.com>
Signed-off-by: David S. Miller <davem@davemloft.net>
Signed-off-by: Jian Liu <jian.liu@windriver.com>
]

3/4 [
Author: Jian Liu
Email: jian.liu@windriver.com
Subject: net: Add variants of capable for use on sockets
Date: Tue, 24 Jun 2014 15:39:27 +0800

commit a3b299da869d6e78cf42ae0b1b41797bcb8c5e4b upstream

sk_net_capable - The common case, operations that are safe in a network namespace.
sk_capable - Operations that are not known to be safe in a network namespace
sk_ns_capable - The general case for special cases.

Signed-off-by: "Eric W. Biederman" <ebiederm@xmission.com>
Signed-off-by: David S. Miller <davem@davemloft.net>
Signed-off-by: Jian Liu <jian.liu@windriver.com>
]

4/4 [
Author: Jian Liu
Email: jian.liu@windriver.com
Subject: net: Add variants of capable for use on netlink messages
Date: Fri, 20 Jun 2014 16:23:34 +0800

commit aa4cf9452f469f16cea8c96283b641b4576d4a7b upstream

netlink_net_capable - The common case use, for operations that are safe on a network namespace
netlink_capable - For operations that are only known to be safe for the global root
netlink_ns_capable - The general case of capable used to handle special cases

__netlink_ns_capable - Same as netlink_ns_capable except taking a netlink_skb_parms instead of
		       the skbuff of a netlink message.

Signed-off-by: "Eric W. Biederman" <ebiederm@xmission.com>
Signed-off-by: David S. Miller <davem@davemloft.net>
Signed-off-by: Jian Liu <jian.liu@windriver.com>
]

Signed-off-by: Yue Tao <yue.tao@windriver.com>