tipc: fix info leaks via msg_name in recv_msg/recv_stream - linux - CVE-2013-3235
Issue: LIN4-25687

1/1 [
Author: Li Wang
Email: li.wang@windriver.com
Subject: tipc: fix info leaks via msg_name in recv_msg/recv_stream
Date: Thu, 26 Jun 2014 14:32:53 +0800

commit 60085c3d009b0df252547adb336d1ccca5ce52ec upstream

The code in set_orig_addr() does not initialize all of the members of
struct sockaddr_tipc when filling the sockaddr info -- namely the union
is only partly filled. This will make recv_msg() and recv_stream() --
the only users of this function -- leak kernel stack memory as the
msg_name member is a local variable in net/socket.c.

Additionally to that both recv_msg() and recv_stream() fail to update
the msg_namelen member to 0 while otherwise returning with 0, i.e.
"success". This is the case for, e.g., non-blocking sockets. This will
lead to a 128 byte kernel stack leak in net/socket.c.

Fix the first issue by initializing the memory of the union with
memset(0). Fix the second one by setting msg_namelen to 0 early as it
will be updated later if we're going to fill the msg_name member.

Cc: Jon Maloy <jon.maloy@ericsson.com>
Cc: Allan Stephens <allan.stephens@windriver.com>
Signed-off-by: Mathias Krause <minipli@googlemail.com>
Signed-off-by: David S. Miller <davem@davemloft.net>
Signed-off-by: Li Wang <li.wang@windriver.com>
]

Signed-off-by: Yue Tao <yue.tao@windriver.com>