writeback: fix pin_sb_for_writeback
Issue: LIN4-31404

1/1 [
Author: Christoph Hellwig
Email: hch@lst.de
Subject: writeback: fix pin_sb_for_writeback
Date: Wed, 9 Jun 2010 15:31:01 +0200

Commit 29cb48594b873f6193d6327097e504bd3e2314de from upstream.

We need to check for s_instances to make sure we don't bother working
against a filesystem that is beeing unmounted, and we need to call
put_super to make sure a superblock is freed when we race against
umount.  Also no need to keep sb_lock after we got a reference on it.

Signed-off-by: Christoph Hellwig <hch@lst.de>
Signed-off-by: Jens Axboe <jaxboe@fusionio.com>
Signed-off-by: Zhang Xiao <xiao.zhang@windriver.com>
]

Signed-off-by: Yue Tao <yue.tao@windriver.com>