jffs2: Fix crash due to truncation of csize
Issue: LIN4-31447

1/1 [
Author: Jian Liu
Email: jian.liu@windriver.com
Subject: jffs2: Fix crash due to truncation of csize
Date: Thu, 26 Jun 2014 17:08:58 +0800

commit 41bf1a24c1001f4d0d41a78e1ac575d2f14789d7 upstream

The variable csize in structure jffs2_tmp_dnode_info is of type uint16_t, but it
is used to hold the compressed data length(csize) which is declared as uint32_t.
So, when the value of csize exceeds 16bits, it gets truncated when assigned to
tn->csize. This is causing a kernel BUG.
Changing the definition of csize in jffs2_tmp_dnode_info to uint32_t fixes the issue.

Signed-off-by: Ajesh Kunhipurayil Vijayan <ajesh@broadcom.com>
Signed-off-by: Kamlakant Patel <kamlakant.patel@broadcom.com>
Signed-off-by: Brian Norris <computersforpeace@gmail.com>
Signed-off-by: Jian Liu <jian.liu@windriver.com>
]

Signed-off-by: Yue Tao <yue.tao@windriver.com>