Add file_ns_capable() helper function for open-time capability checking - linux - CVE-2014-0181
Issue: LIN5-18637

1/5 [
Author: Jian Liu
Email: jian.liu@windriver.com
Subject: Use netlink_ns_capable to verify the permisions of netlink messages
Date: Thu, 8 May 2014 16:20:36 +0800

commit 90f62cf30a78721641e08737bda787552428061e upstream

It is possible by passing a netlink socket to a more privileged
executable and then to fool that executable into writing to the
socket data that happens to be valid netlink message to do
something that privileged executable did not intend to do.
To keep this from happening replace bare capable and
ns_capable calls with netlink_capable, netlink_net_calls
and netlink_ns_capable calls. Which act the same as the
previous calls except they verify that the opener of the
socket had the desired permissions as well.

Reported-by: Andy Lutomirski <luto@amacapital.net>
Signed-off-by: "Eric W. Biederman" <ebiederm@xmission.com>
Signed-off-by: David S. Miller <davem@davemloft.net>
Signed-off-by: Jian Liu <jian.liu@windriver.com>
]

2/5 [
Author: Jian Liu
Email: jian.liu@windriver.com
Subject: netlink: Rename netlink_capable netlink_allowed
Date: Fri, 20 Jun 2014 16:09:44 +0800

commit 5187cd055b6e81fc6526109456f8b20623148d5f upstream

netlink_capable is a static internal function in af_netlink.c and we
have better uses for the name netlink_capable.

Signed-off-by: "Eric W. Biederman" <ebiederm@xmission.com>
Signed-off-by: David S. Miller <davem@davemloft.net>
Signed-off-by: Jian Liu <jian.liu@windriver.com>
]

3/5 [
Author: Jian Liu
Email: jian.liu@windriver.com
Subject: net: Add variants of capable for use on netlink messages
Date: Fri, 20 Jun 2014 16:23:34 +0800

commit aa4cf9452f469f16cea8c96283b641b4576d4a7b upstream

netlink_net_capable - The common case use, for operations that are safe on a network namespace
netlink_capable - For operations that are only known to be safe for the global root
netlink_ns_capable - The general case of capable used to handle special cases

__netlink_ns_capable - Same as netlink_ns_capable except taking a netlink_skb_parms instead of
		       the skbuff of a netlink message.

Signed-off-by: "Eric W. Biederman" <ebiederm@xmission.com>
Signed-off-by: David S. Miller <davem@davemloft.net>
Signed-off-by: Jian Liu <jian.liu@windriver.com>
]

4/5 [
Author: Jian Liu
Email: jian.liu@windriver.com
Subject: net: Add variants of capable for use on on sockets
Date: Fri, 20 Jun 2014 11:09:05 +0800

commit a3b299da869d6e78cf42ae0b1b41797bcb8c5e4b upstream

sk_net_capable - The common case, operations that are safe in a network namespace.
sk_capable - Operations that are not known to be safe in a network namespace
sk_ns_capable - The general case for special cases.

Signed-off-by: "Eric W. Biederman" <ebiederm@xmission.com>
Signed-off-by: David S. Miller <davem@davemloft.net>
Signed-off-by: Jian Liu <jian.liu@windriver.com>
]

5/5 [
Author: Jian Liu
Email: jian.liu@windriver.com
Subject: Add file_ns_capable() helper function for open-time capability checking
Date: Fri, 20 Jun 2014 11:20:28 +0800

commit 935d8aabd4331f47a89c3e1daa5779d23cf244ee upstream

Nothing is using it yet, but this will allow us to delay the open-time
checks to use time, without breaking the normal UNIX permission
semantics where permissions are determined by the opener (and the file
descriptor can then be passed to a different process, or the process can
drop capabilities).

Signed-off-by: Linus Torvalds <torvalds@linux-foundation.org>
Signed-off-by: Jian Liu <jian.liu@windriver.com>
]

Signed-off-by: Yue Tao <yue.tao@windriver.com>