fsl/usb: Workaround for USB erratum-A005697
Issue: LIN5-19007

1/1 [
Author: Zhang Xiao
Email: xiao.zhang@windriver.com
Subject: fsl/usb: Workaround for USB erratum-A005697
Date: Thu, 3 Jul 2014 08:23:27 +0000

Commit c2272ee4e08ef5f0793ff90d3b3c87781438d0ff from
git://git.freescale.com/ppc/sdk/linux.git

As per USB specification, in the Suspend state, the status bit does
not change until the port is suspended. However, there may be a delay
in suspending a port if there is a transaction currently in progress
on the bus.

In the USBDR controller, the PORTSCx[SUSP] bit changes immediately when
the application sets it and not when the port is actually suspended

Workaround for this issue involves waiting for a minimum of 10ms to
allow the controller to go into SUSPEND state before proceeding ahead

Signed-off-by: Ramneek Mehresh <ramneek.mehresh@freescale.com>
Signed-off-by: Nikhil Badola <nikhil.badola@freescale.com>
Change-Id: If27150f914fc675b843e0b40f95908c3da8e2e30
Reviewed-on: http://git.am.freescale.net:8181/13538
Tested-by: Review Code-CDREVIEW <CDREVIEW@freescale.com>
Reviewed-by: Jose Rivera <German.Rivera@freescale.com>
Signed-off-by: Zhang Xiao <xiao.zhang@windriver.com>
]

Signed-off-by: Yue Tao <yue.tao@windriver.com>