ext3: Fix format string issues - linux - CVE-2013-1848
Issue: LIN4-23748

1/1 [
Author: Lars-Peter Clausen
Email: lars@metafoo.de
Subject: ext3: Fix format string issues
Date: Wed, 16 Jul 2014 15:33:43 +0800

commit 8d0c2d10dd72c5292eda7a06231056a4c972e4cc upstream

ext3_msg() takes the printk prefix as the second parameter and the
format string as the third parameter. Two callers of ext3_msg omit the
prefix and pass the format string as the second parameter and the first
parameter to the format string as the third parameter. In both cases
this string comes from an arbitrary source. Which means the string may
contain format string characters, which will
lead to undefined and potentially harmful behavior.

The issue was introduced in commit 4cf46b67eb("ext3: Unify log messages
in ext3") and is fixed by this patch.

CC: stable@vger.kernel.org
Signed-off-by: Lars-Peter Clausen <lars@metafoo.de>
Signed-off-by: Jan Kara <jack@suse.cz>
Signed-off-by: Xufeng Zhang <xufeng.zhang@windriver.com>
]

Signed-off-by: Yue Tao <yue.tao@windriver.com>