aacraid: prevent invalid pointer dereference - linux - CVE-2013-6380
Issue: LIN4-31054

1/1 [
Author: Mahesh Rajashekhara
Email: Mahesh.Rajashekhara@pmcs.com
Subject: aacraid: prevent invalid pointer dereference
Date: Thu, 31 Oct 2013 14:01:02 +0530

commit b4789b8e6be3151a955ade74872822f30e8cd914 upstream

It appears that driver runs into a problem here if fibsize is too small
because we allocate user_srbcmd with fibsize size only but later we
access it until user_srbcmd->sg.count to copy it over to srbcmd.

It is not correct to test (fibsize < sizeof(*user_srbcmd)) because this
structure already includes one sg element and this is not needed for
commands without data.  So, we would recommend to add the following
(instead of test for fibsize == 0).

Signed-off-by: Mahesh Rajashekhara <Mahesh.Rajashekhara@pmcs.com>
Reported-by: Nico Golde <nico@ngolde.de>
Reported-by: Fabian Yamaguchi <fabs@goesec.de>
Signed-off-by: Linus Torvalds <torvalds@linux-foundation.org>
]

Signed-off-by: Yue Tao <yue.tao@windriver.com>