KVM: Improve create VCPU parameter - linux - CVE-2013-4587
Issue: LIN4-31035

1/1 [
Author: Andy Honig
Email: ahonig@google.com
Subject: KVM: Improve create VCPU parameter (CVE-2013-4587)
Date: Mon, 18 Nov 2013 16:09:22 -0800

commit 338c7dbadd2671189cec7faf64c84d01071b3f96 upstream

In multiple functions the vcpu_id is used as an offset into a bitfield.  Ag
malicious user could specify a vcpu_id greater than 255 in order to set or
clear bits in kernel memory.  This could be used to elevate priveges in the
kernel.  This patch verifies that the vcpu_id provided is less than 255.
The api documentation already specifies that the vcpu_id must be less than
max_vcpus, but this is currently not checked.

Reported-by: Andrew Honig <ahonig@google.com>
Cc: stable@vger.kernel.org
Signed-off-by: Andrew Honig <ahonig@google.com>
Signed-off-by: Paolo Bonzini <pbonzini@redhat.com>
]

Signed-off-by: Yue Tao <yue.tao@windriver.com>