ALSA: control: Make sure that id->index does not overflow - linux - CVE-2014-4656
Issue: LIN6-7947

1/2 [
Author: Lars-Peter Clausen
Email: lars@metafoo.de
Subject: ALSA: control: Handle numid overflow
Date: Wed, 18 Jun 2014 13:32:34 +0200

Each control gets automatically assigned its numids when the control is created.
The allocation is done by incrementing the numid by the amount of allocated
numids per allocation. This means that excessive creation and destruction of
controls (e.g. via SNDRV_CTL_IOCTL_ELEM_ADD/REMOVE) can cause the id to
eventually overflow. Currently when this happens for the control that caused the
overflow kctl->id.numid + kctl->count will also over flow causing it to be
smaller than kctl->id.numid. Most of the code assumes that this is something
that can not happen, so we need to make sure that it won't happen

Signed-off-by: Lars-Peter Clausen <lars@metafoo.de>
Acked-by: Jaroslav Kysela <perex@perex.cz>
Cc: <stable@vger.kernel.org>
Signed-off-by: Takashi Iwai <tiwai@suse.de>
]

2/2 [
Author: Lars-Peter Clausen
Email: lars@metafoo.de
Subject: ALSA: control: Make sure that id->index does not overflow
Date: Wed, 18 Jun 2014 13:32:35 +0200

The ALSA control code expects that the range of assigned indices to a control is
continuous and does not overflow. Currently there are no checks to enforce this.
If a control with a overflowing index range is created that control becomes
effectively inaccessible and unremovable since snd_ctl_find_id() will not be
able to find it. This patch adds a check that makes sure that controls with a
overflowing index range can not be created.

Signed-off-by: Lars-Peter Clausen <lars@metafoo.de>
Acked-by: Jaroslav Kysela <perex@perex.cz>
Cc: <stable@vger.kernel.org>
Signed-off-by: Takashi Iwai <tiwai@suse.de>
]

Signed-off-by: Yue Tao <yue.tao@windriver.com>