staging: wlags49_h2: buffer overflow setting station name - linux - CVE-2013-4514
Issue: LIN5-15350

1/1 [
Author: Dan Carpenter
Email: dan.carpenter@oracle.com
Subject: staging: wlags49_h2: buffer overflow setting station name
Date: Thu, 24 Jul 2014 16:46:19 +0800

commit b5e2f339865fb443107e5b10603e53bbc92dc054 upstream

We need to check the length parameter before doing the memcpy().  I've
actually changed it to strlcpy() as well so that it's NUL terminated.

You need CAP_NET_ADMIN to trigger these so it's not the end of the
world.

Reported-by: Nico Golde <nico@ngolde.de>
Reported-by: Fabian Yamaguchi <fabs@goesec.de>
Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>
Cc: stable@kernel.org
Signed-off-by: Linus Torvalds <torvalds@linux-foundation.org>
Signed-off-by: wenlin.kang <wenlin.kang@windriver.com>
]

Signed-off-by: Yue Tao <yue.tao@windriver.com>