usb: dwc3: combine the gadget primary and thread handler for rt kernel
Rally: US40782

1/1 [
Author: Kevin Hao
Email: kexin.hao@windriver.com
Subject: usb: dwc3: combine the gadget primary and thread handler for rt kernel
Date: Fri, 1 Aug 2014 11:02:30 +0800

The dwc3 gadget irq handler is split into two parts. The primary
handler run in irq context, and the thread handler run in thread
context. This works pretty good in standard kernel. But for a rt
kernel this will cause issue due to the spinlock used by these two
parts become sleepable. One option is change this spinlock to raw
spinlock, but this will make the most of the code in thread handler
run with irq disabled. This will cause various issues if the normal
spinlock or work queue is used in the gadget driver callbacks. So
we choose to combine the gadget interrupt into one function and make
it entirely threaded for a rt kernel.

Signed-off-by: Kevin Hao <kexin.hao@windriver.com>
]

Signed-off-by: Yue Tao <yue.tao@windriver.com>