ipv6: add support of peer address
Issue: LIN5-19139

1/1 [
Author: yzhu1
Email: yanjun.zhu@windriver.com
Subject: ipv6: add support of peer address
Date: Thu, 31 Jul 2014 23:59:56 -0700

commit caeaba79009c2ee858c3b2bf8caf922cd719fead upstream

This patch adds the support of peer address for IPv6. For example, it is
possible to specify the remote end of a 6inY tunnel.
This was already possible in IPv4:
ip addr add ip1 peer ip2 dev dev1

The peer address is specified with IFA_ADDRESS and the local address with
IFA_LOCAL (like explained in include/uapi/linux/if_addr.h).
Note that the API is not changed, because before this patch, it was not
possible to specify two different addresses in IFA_LOCAL and IFA_REMOTE.
There is a small change for the dump: if the peer is different from ::,
IFA_ADDRESS will contain the peer address instead of the local address.

Signed-off-by: Nicolas Dichtel <nicolas.dichtel@6wind.com>
Signed-off-by: David S. Miller <davem@davemloft.net>
Signed-off-by: yzhu1 <yanjun.zhu@windriver.com>
]

Signed-off-by: Yue Tao <yue.tao@windriver.com>