AUFS: ensure module is build for all OVP loads.
Change the cfg setting from =y to =m.  Also, since we were
getting to AUFS via an include of "... /standard.scc nocfg"
the AUFS config fragment was being ignored.  So here we do
explicitly include it again, but with "nopatch" so all we
get is the config fragment.

Note that the other settings visible in the context are
still set to =y because they are bool tuning settings
and not tristate; hence indifferent to whether AUFS_FS=m

Signed-off-by: Paul Gortmaker <paul.gortmaker@windriver.com>