stable: integrate 49 --> 53 as individual patches.
There are four patches to rtmutex code in 3.10.49 that are disruptive
to the -rt codebase.  As they have not been merged into the stable-rt
tree yet, we can not look there for a merge solution.  So for now, we
make the switch to individual patches to get all the rest of the
stable content as an interim solution.

Signed-off-by: Paul Gortmaker <paul.gortmaker@windriver.com>