stable: replace individual patch-by-patch stable content with reverts
We used patch-by-patch instances of 49 (and following releases) to get
rid of the conflicting rtmutex patches that don't play nice with -rt.

A cleaner way that allows us to keep Greg's tags and Makefile version
commits is to just pre-treegen revert the four offending changes.

The resulting source trees are identical.

We bump kver to match the additional 5 releases that previously existed
as patch dirs.

Signed-off-by: Paul Gortmaker <paul.gortmaker@windriver.com>