meta: intel-common-standard: add ericson-3g to intel-common-standard.scc
BSPs that use intel-common-standard.scc should be expected to work on some
devices with integrated 3G modems, most notably the FRI2. However, because
we only have a configuration fragment for one 3g driver, it's not worth
creating a 3g-all.scc file. For now, the best option is to add that driver
directly to intel-common-standard.scc.

Signed-off-by: Max Eliaser <max.eliaser@intel.com>
Signed-off-by: Bruce Ashfield <bruce.ashfield@windriver.com>