SUNRPC: track whether a request is coming from a loop-back interface.
1/2 [
Author: NeilBrown
Email: neilb@suse.de
Subject: SUNRPC: track whether a request is coming from a loop-back interface.
Date: Mon, 18 Aug 2014 13:51:07 -0700

commit ef11ce24875a8a540adc185e7bce3d7d49c8296f upstream

If an incoming NFS request is coming from the local host, then
nfsd will need to perform some special handling.  So detect that
possibility and make the source visible in rq_local.

Signed-off-by: NeilBrown <neilb@suse.de>
Signed-off-by: J. Bruce Fields <bfields@redhat.com>
Signed-off-by: Yang Shi <yang.shi@windriver.com>
Signed-off-by: Bruce Ashfield <bruce.ashfield@windriver.com>
]

2/2 [
Author: NeilBrown
Email: neilb@suse.de
Subject: nfsd: Only set PF_LESS_THROTTLE when really needed.
Date: Mon, 18 Aug 2014 13:51:08 -0700

commit 8658452e4a588da603f6cb5ee2615deafcd82b71 upstream

PF_LESS_THROTTLE has a very specific use case: to avoid deadlocks
and live-locks while writing to the page cache in a loop-back
NFS mount situation.

It therefore makes sense to *only* set PF_LESS_THROTTLE in this
situation.
We now know when a request came from the local-host so it could be a
loop-back mount.  We already know when we are handling write requests,
and when we are doing anything else.

So combine those two to allow nfsd to still be throttled (like any
other process) in every situation except when it is known to be
problematic.

Signed-off-by: NeilBrown <neilb@suse.de>
Signed-off-by: J. Bruce Fields <bfields@redhat.com>
Signed-off-by: Yang Shi <yang.shi@windriver.com>
Signed-off-by: Bruce Ashfield <bruce.ashfield@windriver.com>
]

Signed-off-by: Bruce Ashfield <bruce.ashfield@windriver.com>