gre: add x-netns support
1/5 [
Author: Nicolas Dichtel
Email: nicolas.dichtel@6wind.com
Subject: gre: add x-netns support
Date: Wed, 27 Aug 2014 08:51:32 +0800

commit b57708add31494175be741ed3fd24023b50c3423 upstream

This patch allows to switch the netns when packet is encapsulated or
decapsulated. In other word, the encapsulated packet is received in a netns,
where the lookup is done to find the tunnel. Once the tunnel is found, the
packet is decapsulated and injecting into the corresponding interface which
stands to another netns.

When one of the two netns is removed, the tunnel is destroyed.

Signed-off-by: Nicolas Dichtel <nicolas.dichtel@6wind.com>
Signed-off-by: David S. Miller <davem@davemloft.net>
Signed-off-by: Jianchuan Wang <jianchuan.wang@windriver.com>
Signed-off-by: Bruce Ashfield <bruce.ashfield@windriver.com>
]

2/5 [
Author: stephen hemminger
Email: stephen@networkplumber.org
Subject: gre: allow changing mac address when device is up
Date: Wed, 27 Aug 2014 08:51:33 +0800

commit f8c1b7ce00254a5bb75d5b5e5ef1601326a0e08e upstream

There is no need to require forcing device down on a Ethernet GRE (gretap)
tunnel to change the MAC address.

Signed-off-by: Stephen Hemminger <stephen@networkplumber.org>
Signed-off-by: David S. Miller <davem@davemloft.net>
Signed-off-by: Jianchuan Wang <jianchuan.wang@windriver.com>
Signed-off-by: Bruce Ashfield <bruce.ashfield@windriver.com>
]

3/5 [
Author: Nicolas Dichtel
Email: nicolas.dichtel@6wind.com
Subject: ip6gre: add x-netns support
Date: Wed, 27 Aug 2014 08:51:34 +0800

commit 22f08069e8b415b827e910ad75ed55eeadc4a877 upstream

This patch allows to switch the netns when packet is encapsulated or
decapsulated. In other word, the encapsulated packet is received in a netns,
where the lookup is done to find the tunnel. Once the tunnel is found, the
packet is decapsulated and injecting into the corresponding interface which
stands to another netns.

When one of the two netns is removed, the tunnel is destroyed.

Signed-off-by: Nicolas Dichtel <nicolas.dichtel@6wind.com>
Signed-off-by: David S. Miller <davem@davemloft.net>
Signed-off-by: Jianchuan Wang <jianchuan.wang@windriver.com>
Signed-off-by: Bruce Ashfield <bruce.ashfield@windriver.com>
]

4/5 [
Author: Tom Herbert
Email: therbert@google.com
Subject: net: Generalize checksum_init functions
Date: Wed, 27 Aug 2014 08:51:35 +0800

commit 76ba0aae673075c77a8b775e9133c8e8b1a44563 upstream

Create a general __skb_checksum_validate function (actually a
macro) to subsume the various checksum_init functions. This
function can either init the checksum, or do the full validation
(logically checksum_init+skb_check_complete)-- a flag specifies
if full vaidation is performed. Also, there is a flag to the function
to indicate that zero checksums are allowed (to support optional
UDP checksums).

Added several stub functions for calling __skb_checksum_validate.

Signed-off-by: Tom Herbert <therbert@google.com>
Signed-off-by: David S. Miller <davem@davemloft.net>
Signed-off-by: Jianchuan Wang <jianchuan.wang@windriver.com>
Signed-off-by: Bruce Ashfield <bruce.ashfield@windriver.com>
]

5/5 [
Author: Tom Herbert
Email: therbert@google.com
Subject: gre6: Call skb_checksum_simple_validate
Date: Wed, 27 Aug 2014 08:51:36 +0800

commit 81249bea1fb003b6fcbc1709dd5a5fc1e26e168d upstream

Use skb_checksum_simple_validate to verify checksum.

Signed-off-by: Tom Herbert <therbert@google.com>
Signed-off-by: David S. Miller <davem@davemloft.net>
Signed-off-by: Jianchuan Wang <jianchuan.wang@windriver.com>
Signed-off-by: Bruce Ashfield <bruce.ashfield@windriver.com>
]

Signed-off-by: Bruce Ashfield <bruce.ashfield@windriver.com>