ALSA: control: Handle numid overflow - linux - CVE-2014-4656
Issue: LIN4-31570

1/4 [
Author: Lars-Peter Clausen
Email: lars@metafoo.de
Subject: ALSA: control: Make sure that id->index does not overflow
Date: Wed, 18 Jun 2014 13:32:35 +0200

Commit 883a1d49f0d77d30012f114b2e19fc141beb3e8e upstream

The ALSA control code expects that the range of assigned indices to a control is
continuous and does not overflow. Currently there are no checks to enforce this.
If a control with a overflowing index range is created that control becomes
effectively inaccessible and unremovable since snd_ctl_find_id() will not be
able to find it. This patch adds a check that makes sure that controls with a
overflowing index range can not be created.

Signed-off-by: Lars-Peter Clausen <lars@metafoo.de>
Acked-by: Jaroslav Kysela <perex@perex.cz>
Cc: <stable@vger.kernel.org>
Signed-off-by: Takashi Iwai <tiwai@suse.de>
Signed-off-by: Yue Tao <Yue.Tao@windriver.com>
]

2/4 [
Author: Clemens Ladisch
Email: clemens@ladisch.de
Subject: ALSA: control: fix numid conflict check for new controls
Date: Mon, 7 Mar 2011 13:22:50 +0100

Commit 7c7335877105364f7f5181e80ff34206b54be81f upstream

The purpose of the snd_ctl_hole_check() function is to find conflicts
between the numerical IDs of the new control and those of any existing
controls.  However, it would fail to detect an existing control whose
count is smaller than the new control's count and whose interval of IDs
is entirely contained in the interval of the new control's IDs.

To fix this, use the correct formula to detect overlapping intervals,
which happens to simplify the condition.

This problem was not encountered so far because ALSA does not yet allow
drivers to allocate specific control IDs.

Signed-off-by: Clemens Ladisch <clemens@ladisch.de>
Signed-off-by: Takashi Iwai <tiwai@suse.de>
]

3/4 [
Author: Clemens Ladisch
Email: clemens@ladisch.de
Subject: ALSA: control: clean up snd_ctl_hole_check()
Date: Mon, 7 Mar 2011 13:24:30 +0100

Commit 0e82e5fa97614c9ca3efca5f8dca69dffd1c0ec0 upstream

The return value of snd_ctl_hole_check() is used only to detect whether
to continue the loop in snd_ctl_find_hole() or not, so we can simplify
the code by changing this return type to a boolean.  Also rename this
function to better show what it actually does.

Signed-off-by: Clemens Ladisch <clemens@ladisch.de>
Signed-off-by: Takashi Iwai <tiwai@suse.de>
]

4/4 [
Author: Lars-Peter Clausen
Email: lars@metafoo.de
Subject: ALSA: control: Handle numid overflow
Date: Wed, 18 Jun 2014 13:32:34 +0200

Commit ac902c112d90a89e59916f751c2745f4dbdbb4bd upstream

Each control gets automatically assigned its numids when the control is created.
The allocation is done by incrementing the numid by the amount of allocated
numids per allocation. This means that excessive creation and destruction of
controls (e.g. via SNDRV_CTL_IOCTL_ELEM_ADD/REMOVE) can cause the id to
eventually overflow. Currently when this happens for the control that caused the
overflow kctl->id.numid + kctl->count will also over flow causing it to be
smaller than kctl->id.numid. Most of the code assumes that this is something
that can not happen, so we need to make sure that it won't happen

Signed-off-by: Lars-Peter Clausen <lars@metafoo.de>
Acked-by: Jaroslav Kysela <perex@perex.cz>
Cc: <stable@vger.kernel.org>
Signed-off-by: Takashi Iwai <tiwai@suse.de>
]

Signed-off-by: Yue Tao <yue.tao@windriver.com>