rt: update name of localversion-rt patch to match version [README!]
At this point a comment is warranted about the prev. commits.  We had
a revert of stable mutex changes that allowed us to continue using
stable with -rt while waiting for RT to overlap the same changes.

As that has since happened, we've dropped the reverts and updated to
the latest -rt.  However, due to the complexity, the -rt maintainers
only did a ff merge of stable into their -rt (no rebased commits to
import).  Hence I'd created locally a rebased version for tags from
48 --> 56 -rt and then diff'd those trees against the maintanier -rt
tags of the same (to ensure 100% source equivalence) and then we've
imported those rt series of rebased patches here into the kernel-cache
one by one in the previous commits.

Signed-off-by: Paul Gortmaker <paul.gortmaker@windriver.com>