intel: create intel-common-preempt-rt and use it
The -preempt variants were including intel-common-standard.scc, and
then the preempt-rt kernel type.

But since preempt-rt also includes the standard kernel, this results
in a double include of the standard kernel, at a different point in
the branch structure, and hence a less than obvious branch is created
like: standard/standard/preempt-rt.

To fix this, we create a parallel intel-common-preempt-rt and include
it when necessary.

Signed-off-by: Bruce Ashfield <bruce.ashfield@windriver.com>