meta: remove "net: Rename skb->rxhash to skb->hash"
This is used to undo the previous commit "net: Rename skb->rxhash to skb->hash"
which introduces unnecessary renamings for BPF. Build will fail if related
drivers are enabled but do not get updated accordingly. Since the context has
been changed by other commits in the BPF set, some fixes need to be done as
follow.

 - Renaming for net/core/filter.c has been removed by "net: filter:
   rework/optimize internal BPF interpreter's instruction set"
   Nothing to undo

 - Renaming for arch/x86/net/bpf_jit_comp.c has been removed by "net: filter:
   x86: internal BPF JIT"
   Nothing to undo

 - Some more references to "hash" in net/core/filter.c are introduced by "net:
   filter: make BPF conversion more readable"
   Rename hash to rxhash

Signed-off-by: He Zhe <zhe.he@windriver.com>
Signed-off-by: Bruce Ashfield <bruce.ashfield@windriver.com>