PM / hibernate: introduce "nohibernate" boot parameter
1/1 [
Author: Kees Cook
Email: keescook@chromium.org
Subject: PM / hibernate: introduce "nohibernate" boot parameter
Date: Fri, 13 Jun 2014 13:30:35 -0700

commit a6e15a39048ec3229b9a53425f4384f55f6cc1b3 upstream.

To support using kernel features that are not compatible with hibernation,
this creates the "nohibernate" kernel boot parameter to disable both
hibernation and resume. This allows hibernation support to be a boot-time
choice instead of only a compile-time choice.

Signed-off-by: Kees Cook <keescook@chromium.org>
Acked-by: Pavel Machek <pavel@ucw.cz>
Signed-off-by: Rafael J. Wysocki <rafael.j.wysocki@intel.com>
[PG: Fixes yocto-3.14 6cfe9c448741ab, which depends on nohibernate.]
Signed-off-by: Paul Gortmaker <paul.gortmaker@windriver.com>
]

Signed-off-by: Bruce Ashfield <bruce.ashfield@windriver.com>