vxlan: re-add the patch based on the original
The backport had a lot of added CONFIG_VXLAN added to what was the
mainline code that had nothing similar.  This was OK for configs
where it was enabled, but is problematic for modular configs where
the boolean nature of #ifdef CONFIG_VXLAN fails.

Switch back to the unmolested version of the mainline commit that
has no such #ifdef code added to it, since we need the struct
fields and inline functions for such modular builds.

Signed-off-by: Paul Gortmaker <paul.gortmaker@windriver.com>