cryptodev: stomp dynamic version numbering for in tree builds
1/1 [
Author: Paul Gortmaker
Email: paul.gortmaker@windriver.com
Subject: cryptodev: stomp dynamic version numbering for in tree builds
Date: Sat, 27 Sep 2014 11:49:34 -0400

The temporary creation of a version header adds complications
for builds where src != build (of which it doesn't use $(src)
at all) and plus the dependencies on version.h were not working,
leading to compile failures.

For an in tree build, the version will be never changing, so
just make a static version header and junk all the Makefile
complexity that we don't need.

Signed-off-by: Paul Gortmaker <paul.gortmaker@windriver.com>
]

Signed-off-by: Bruce Ashfield <bruce.ashfield@windriver.com>