lttng2: added lttng Kconfig for in kernel building.
1/2 [
Author: Jianchuan Wang
Email: jianchuan.wang@windriver.com
Subject: lttng2: added lttng Kconfig for in kernel building.
Date: Fri, 19 Sep 2014 09:50:56 +0800

The default lttng-modules tree is meant to be compiled as an external kernel
module and does not have a Kconfig.  The LTSI kernel added lttng 2.0 to their
kernel tree and this Kconfig is taken from:
http://git.linuxfoundation.org/?p=ltsi-kernel.git;a=blob;f=patches.lttng/0017-lttng-toplevel-Makefile-and-Kconfig.patch
The changes made from the above:
	-removed Makefile, use the newer one from lttng-modules drop
	-change install directory to lttng2
	-change Kconfig to LTTNG2 to avoid collisions with previous versions
	-enforce compiling as a kernel module
	-make sure the old LTTNG is not defined if configuring for LTTNG2

Note: this driver must be compiled as a kernel module.  The lttng-tools
expects to modprobe all the lttng modules and fails otherwise.
Signed-off-by: Jianchuan Wang <jianchuan.wang@windriver.com>
]

2/2 [
Author: Jianchuan Wang
Email: jianchuan.wang@windriver.com
Subject: lttng2: enable the lttng driver in the build
Date: Fri, 19 Sep 2014 09:54:03 +0800

Enable the build for the lttng module in drivers/stagging/lttng2.

Signed-off-by: Jianchuan Wang <jianchuan.wang@windriver.com>
]

Signed-off-by: Yang Shi <yang.shi@windriver.com>