net: fsl: make ucc_geth depend on PPC
Issue: LIN7-1288,LIN7-1235,LIN7-1236

1/3 [
Author: Kevin Hao
Email: kexin.hao@windriver.com
Subject: dma: make fsldma depend on PPC
Date: Tue, 28 Oct 2014 13:56:40 +0800

The fsldma driver is only buildable on ppc platform. Previously it has
a dependency on FSL_SOC. This was fine since FSL_SOC was only enabled
on some ppc boards. But with the introduction of fsl ls boards, the
FSL_SOC may be enabled on a arm platform and would cause build error
if this driver is enabled. So make it also depend on PPC.

Signed-off-by: Kevin Hao <kexin.hao@windriver.com>
]

2/3 [
Author: Kevin Hao
Email: kexin.hao@windriver.com
Subject: crypto: make talitos depends on PPC
Date: Tue, 28 Oct 2014 14:10:52 +0800

The talitos driver is only buildable on ppc platform. Previously it has
a dependency on FSL_SOC. This was fine since FSL_SOC was only enabled
on some ppc boards. But with the introduction of fsl ls boards, the
FSL_SOC may be enabled on a arm platform and would cause build error
if this driver is enabled. So make it also depend on PPC.

Signed-off-by: Kevin Hao <kexin.hao@windriver.com>
]

3/3 [
Author: Kevin Hao
Email: kexin.hao@windriver.com
Subject: net: fsl: make ucc_geth depend on PPC
Date: Tue, 28 Oct 2014 14:23:20 +0800

The ucc_geth driver is only buildable on ppc platform. Previously it has
a dependency on QUICC_ENGINE. This was fine since QUICC_ENGINE was only
enabled on some ppc boards. But with the introduction of fsl ls boards,
the QUICC_ENGINE may be enabled on a arm platform and would cause build
error if this driver is enabled. So make it also depend on PPC.

Signed-off-by: Kevin Hao <kexin.hao@windriver.com>
]

Signed-off-by: Yue Tao <yue.tao@windriver.com>